HBASE-23811. [OpenTracing] Add shaded JaegerTracing tracer to hbase-thirdparty.
This commit adds a new artifact org.hbase.thirdparty:hbase-shaded-jaeger.
It shades io.jagertracing:jaeger-client and its dependencies, relocating
the classpath. In particular, it requires libthrift 0.13.0 which is
incompatible with libthrift 0.12.0 shipped by the current HBase.

Dependencies that are already included by hbase-common are excluded from
the uber jar. The resulting jar is about 3.9 MB in size.