HBASE-24858 Exclude the current modules jar from the shading step
When a module's jar from the previous build of that module is left in
place, replacing the main artifact for that module, the next build of
that module will happen to pick up the shaded jar and try to re-shaded
that jar.

This is unexpected, as we're reading through all of the module's
dependencies and unpackaging/relocating each of them. We do not expect
to be getting them from the modules main artifact.

The easy solution is that we just always exclude this modules jar. Given
that these modules have no source (just a repacking of thirdparty
projects), this is safe for us to do. Best as I can tell, the resulting
jars are equivalently built.